Split Coordinator and Queue
